feat(session-replay-react-native): refactor package and adopt tsx
This commit refactors the session-replay-react-native package.

Key changes include:
- Renaming the Android package from
  'com.amplitude.pluginsessionreplayreactnative' to
  'com.amplitude.sessionreplayreactnative' for simplicity and consistency.
- Converting TypeScript (.ts) files to TSX (.tsx), allowing the use of JSX
  for React Native components. This includes renaming 'app-mask-view' to
  'amp-mask-view' to align with Amplitude's branding.
- Updating build configurations and manifest files to reflect these changes.

These changes align the package with modern React Native development practices
and improve maintainability.
