+++ packages/session-replay-react-native/android/gradle.properties
@@ -1,5 +1,5 @@
-PluginSessionReplayReactNative_kotlinVersion=1.7.0
-PluginSessionReplayReactNative_minSdkVersion=21
-PluginSessionReplayReactNative_targetSdkVersion=31
-PluginSessionReplayReactNative_compileSdkVersion=31
-PluginSessionReplayReactNative_ndkversion=21.4.7075529
+SessionReplayReactNative_kotlinVersion=1.7.0
+SessionReplayReactNative_minSdkVersion=21
+SessionReplayReactNative_targetSdkVersion=31
+SessionReplayReactNative_compileSdkVersion=31
+SessionReplayReactNative_ndkversion=21.4.7075529
