+++ packages/session-replay-react-native/android/src/main/java/com/amplitude/sessionreplayreactnative/SessionReplayReactNativePackage.kt
@@ -1,4 +1,4 @@
-package com.amplitude.pluginsessionreplayreactnative
+package com.amplitude.sessionreplayreactnative
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
@@ -6,12 +6,12 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 
 
-class PluginSessionReplayReactNativePackage : ReactPackage {
+class SessionReplayReactNativePackage : ReactPackage {
   override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-    return listOf(PluginSessionReplayReactNativeModule(reactContext))
+    return listOf(SessionReplayReactNativeModule(reactContext))
   }
 
   override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
-    return listOf(PluginSessionReplayViewManager())
+    return listOf(SessionReplayReactNativeViewManager())
   }
 }
