+++ packages/session-replay-react-native/src/amp-mask-view.tsx
